refactor: PR 반영
- LottoMachine(LottoStore) 의 과한 책임 분산으로 Winner 클래스 분리
- 분리에 따른 테스트 작성
